requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.2
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
environ==1.0
factory-boy==3.3.0
Faker==23.1.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.3
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.0
